fix another review comment (LongParemterList removal)
