+++ dataland-community-manager/src/main/kotlin/org/dataland/datalandcommunitymanager/services/DataRequestQueryManager.kt
@@ -11,7 +11,6 @@ import org.dataland.datalandcommunitymanager.entities.CompanyRoleAssignmentEntit
 import org.dataland.datalandcommunitymanager.entities.DataRequestEntity
 import org.dataland.datalandcommunitymanager.exceptions.DataRequestNotFoundApiException
 import org.dataland.datalandcommunitymanager.model.companyRoles.CompanyRole
-import org.dataland.datalandcommunitymanager.model.dataRequest.AccessStatus
 import org.dataland.datalandcommunitymanager.model.dataRequest.AggregatedDataRequest
 import org.dataland.datalandcommunitymanager.model.dataRequest.ExtendedStoredDataRequest
 import org.dataland.datalandcommunitymanager.model.dataRequest.RequestStatus
@@ -149,25 +148,11 @@ class DataRequestQueryManager(
      * @param datalandCompanyId the Dataland company ID to apply to the data request
      * @return all filtered data requests
      */
-    @Suppress("LongParameterList") // TODO maybe remove later or look for other workaround
     @Transactional
     fun getDataRequests(
-        dataType: DataTypeEnum?,
-        userId: String?,
-        requestStatus: RequestStatus?,
-        accessStatus: AccessStatus?,
-        reportingPeriod: String?,
-        datalandCompanyId: String?,
+        filter: DataRequestsQueryFilter,
         companyRoleAssignmentsOfCurrentUser: List<CompanyRoleAssignmentEntity>,
     ): List<StoredDataRequest>? {
-        val filter = DataRequestsQueryFilter(
-            dataTypeFilter = dataType?.value ?: "",
-            userIdFilter = userId ?: "",
-            requestStatus = requestStatus,
-            accessStatus = accessStatus,
-            reportingPeriodFilter = reportingPeriod ?: "",
-            datalandCompanyIdFilter = datalandCompanyId ?: "",
-        )
         return createStoredDataRequestObjects(filter, companyRoleAssignmentsOfCurrentUser)
     }
